# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Chronicle API
  version: 1.0.0
  description: REST API for Chronicle, a self-hosted TTRPG worldbuilding platform. Provides programmatic access to campaign data for external tool integration (Foundry VTT, custom scripts).
servers:
  - url: "{baseUrl}/api/v1"
    variables:
      baseUrl:
        default: http://localhost:8080
        description: Chronicle server base URL
security:
  - BearerApiKey: []
...
paths:
...
  /campaigns/{campaignId}:
    get:
      tags: [Campaigns]
      summary: Get campaign details
      operationId: getCampaign
      parameters:
        - $ref: "#/components/parameters/campaignId"
      responses:
        "200":
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"

  # --- Entity Types ---
...
  /campaigns/{campaignId}/calendar/export:
    get:
      tags: [Calendar]
      summary: Export calendar JSON
      operationId: exportCalendar
      parameters:
        - $ref: "#/components/parameters/campaignId"
        - name: events
          in: query
          description: Include events in export
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        "200":
          description: Calendar export data
          content:
            application/json:
              schema:
                type: object
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    BearerApiKey:
      type: http
      scheme: bearer
      description: >
        API key passed as Bearer token. Keys are scoped to a single campaign
        with permission levels: read, write, sync. Rate limited per key.
  parameters:
    campaignId:
      name: campaignId
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        code:
          type: integer
        type:
          type: string
        message:
          type: string
...
    Campaign:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        is_public:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # --- Entities ---
...